requirements.txt
numpy
pandas
matplotlib
scikit-learn>=1.1
Pillow
imageio
glob2
torch